gcp_nlp_agentic_homeassignment_Google_SDK/requirements.txt
google-cloud-language==2.13.4
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.69.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.5
sumy==0.11.0
spacy==3.7.5
# Optional / dev
pytest==8.2.2
matplotlib==3.8.4
